python/requirements.txt
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-OAuth @ git+https://github.com/mitsuhiko/flask-oauth@af6432621a9add5e55b0d7638e1380f66f3d4ed0
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauth2==1.9.0.post1
python-dotenv==0.15.0
requests==2.25.0
urllib3==1.26.2
Werkzeug==0.16.1
